requirements.txt
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
